+++ src/scripts/components/PopupWithImage.js
