+++ src/scripts/components/UserInfo.js
